[v2] leave the binary if testing a single file (useful for anatomization)
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@4219 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ v2/output_tests/test_output.sh
@@ -35,7 +35,9 @@ test_file() {
   strip "Running under ThreadSanitizer"
   strip "========="
   FileCheck < test.out $SRC
-  rm -f a.out test.out *.tmp *.tmp2
+  if [ "$3" == "" ]; then
+    rm -f a.out test.out *.tmp *.tmp2
+  fi
 }
 
 if [ "$1" == "" ]; then
